add route tests for get listing with applicants
